refactor: organise audio functions into separate files
